fix 현재 모집 취업처 리스트 컴포넌트화
