+++ components/companyList/companyList.module.css
@@ -1,0 +1,5 @@
+.wrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+}
