+++ components/companyList/index.tsx
@@ -1,5 +1,12 @@
+import { CurrentRecruitmentCompanyList } from "../common/CurrentRecruitmentCompanyList";
+import s from "./companyList.module.css";
+
 const CompanyList = () => {
-  return <>기업 목록</>;
+  return (
+    <div className={s.wrapper}>
+      <CurrentRecruitmentCompanyList />
+    </div>
+  );
 };
 
 export default CompanyList;
