+++ components/main/index.tsx
@@ -1,28 +1,10 @@
 import s from "./main.module.css";
 import ProfileCard from "../common/ProfileCard";
-import CompanyCardList from "../common/CompanyCardList";
 import SlidePage from "./slidePage";
 import Chart from "./Chart";
-import recruitment from "../../apis/company/recruitment";
-import { useQuery } from "react-query";
-import { QueryKeys } from "../../constants/queryKeys";
-import { useRecoilValue } from "recoil";
-import { searchRequirementState } from "./../../atoms/searchRequirement";
-import CompanyListFilter from "../common/CompanyListFilter";
-import { LoadingSpiner } from "../common/LoadingSpiner";
-import ListTitle from "../common/ListTitle";
+import { CurrentRecruitmentCompanyList } from "../common/CurrentRecruitmentCompanyList";
 
 const Main = () => {
-  const searchRequirement = useRecoilValue(searchRequirementState);
-
-  const currentRecruimentCompanyQuery = useQuery(
-    [QueryKeys.currentRecruitmentCompanyList, searchRequirement],
-    () => recruitment.getCurrentRecruitmentCompanyList(searchRequirement),
-    {
-      staleTime: Infinity,
-    }
-  );
-
   return (
     <div className={s.wrapper}>
       <div className={s.grid_wrapper}>
@@ -34,15 +16,7 @@ const Main = () => {
         <ProfileCard />
         <Chart />
       </div>
-      <CompanyListFilter />
-      <ListTitle content="현재 모집중인 취업처" />
-      {currentRecruimentCompanyQuery.isLoading ? (
-        <LoadingSpiner />
-      ) : (
-        <CompanyCardList
-          companyList={currentRecruimentCompanyQuery.data?.data.recruitments}
-        />
-      )}
+      <CurrentRecruitmentCompanyList />
     </div>
   );
 };
